Make recommended changes
