+++ src/mobile/src/libs/store.js
@@ -1,6 +1,7 @@
 import get from 'lodash/get';
 import { getVersion, getBuildNumber } from 'react-native-device-info';
 import { AsyncStorage } from 'react-native';
+import { doesSaltExistInKeychain } from 'libs/keychain';
 import store, { persistStore, purgeStoredState, createPersistor } from '../../../shared/store';
 import initializeApp from '../ui/routes/entry';
 import { setAppVersions, resetWallet } from '../../../shared/actions/settings';
@@ -21,10 +22,36 @@ const shouldMigrate = (restoredState) => {
     return restoredVersion !== currentVersion || restoredBuildNumber !== currentBuildNumber;
 };
 
+const resetIfKeychainIsEmpty = (state) => {
+    doesSaltExistInKeychain().then((exists) => {
+        if (!exists) {
+            purgeStoredState({ storage: persistConfig.storage })
+                .then(() => {
+                    state.dispatch(resetWallet());
+                    // Set the new app version
+                    state.dispatch(
+                        setAppVersions({
+                            version: getVersion(),
+                            buildNumber: getBuildNumber(),
+                        }),
+                    );
+                    return initializeApp(state);
+                })
+                .catch((err) => console.error(err)); // eslint-disable-line no-console
+        }
+    });
+};
+
 const migrate = (state, restoredState) => {
     // TODO: Doing a dirty patch to disable migration setup for alpha v0.2.0
     // since this would be installed as a fresh application.
     const hasAnUpdate = shouldMigrate(restoredState);
+
+    // Reset wallet if keychain is empty
+    // Fixes issues related to iCloud backup
+    if (state.accounts.onboardingComplete) {
+        resetIfKeychainIsEmpty(state);
+    }
 
     if (!hasAnUpdate) {
         state.dispatch(
